Remove misleading comment about script property
The `script` property is fit to hold secrets if it is defined under
`protectedSettings`.

The phrase "instead of `script`" is the result of a mistaken
clarification.

The original comment was "Use this property instead if your command
contains secrets such as passwords" and it meant one should use the
"protected" (defined under `protectedSettings`) `commandToExecute`
instead of the "public" (defined under `protectedSettings`)
`commandToExecute` if the command contains secrets.

After the mutually exclusive `script` property was added, and the
list of properties (now three) that could be either protected or
public was put into its own section, that comment no longer made
sense. So "instead" was changed to "instead of `script`" apparently
in attempt to make sense of it, but this is also incorrect. The
whole sentence should just be deleted.